Update README.md
The application allows each user to find the dog park closest to his location by using the user's geolocation. Upon entering the application, the user is required to enter a username and password, if he is not registered (does not appear in the DB) he is transferred to a registration page where he enters information about himself and his dog. After From here he is transferred to a page that allows him to choose the size of the dog park he is looking for (large / small) and gives permission to use his location. In the end, the customer will see a detailed table that includes all the dog parks near his location and information about the park including its size and distance from his location. During the use of the application The customer will come across information about what it means to be a member of the application and the various benefits we provide.